ISPN-5029 Make FileLookup backwards compatible with 6.x
Conflicts:
	commons/src/main/java/org/infinispan/commons/util/FileLookup.java
	core/src/test/java/org/infinispan/configuration/ConfigurationUnitTest.java
