+++ core/src/test/java/org/infinispan/configuration/ConfigurationUnitTest.java
@@ -15,6 +15,7 @@ import javax.xml.validation.SchemaFactory;
 import org.infinispan.Cache;
 import org.infinispan.commons.CacheConfigurationException;
 import org.infinispan.commons.util.FileLookup;
+import org.infinispan.commons.util.FileLookupFactory;
 import org.infinispan.configuration.cache.CacheMode;
 import org.infinispan.configuration.cache.Configuration;
 import org.infinispan.configuration.cache.ConfigurationBuilder;
@@ -171,7 +172,7 @@ public class ConfigurationUnitTest extends AbstractInfinispanTest {
 
    @Test
    public void testSchema() throws Exception {
-      FileLookup lookup = new FileLookup();
+      FileLookup lookup = FileLookupFactory.newInstance();
       URL schemaFile = lookup.lookupFileLocation("schema/infinispan-config-7.0.xsd", Thread.currentThread().getContextClassLoader());
       Source xmlFile = new StreamSource(lookup.lookupFile("configs/unified/all.xml", Thread.currentThread().getContextClassLoader()));
       SchemaFactory.newInstance(XMLConstants.W3C_XML_SCHEMA_NS_URI).newSchema(schemaFile).newValidator().validate(xmlFile);
