+++ core/src/test/java/org/infinispan/distexec/mapreduce/BaseLargeWordCountMapReduceTest.java
@@ -13,6 +13,7 @@ import java.util.StringTokenizer;
 
 import org.infinispan.Cache;
 import org.infinispan.commons.util.FileLookup;
+import org.infinispan.commons.util.FileLookupFactory;
 import org.infinispan.configuration.cache.CacheMode;
 import org.infinispan.configuration.cache.ConfigurationBuilder;
 import org.infinispan.test.MultipleCacheManagersTest;
@@ -117,7 +118,7 @@ public abstract class BaseLargeWordCountMapReduceTest extends MultipleCacheManag
 
       Cache c1 = cache(0, cacheName());
       Cache c2 = cache(1, cacheName());
-      FileLookup fileLookup = new FileLookup();
+      FileLookup fileLookup = FileLookupFactory.newInstance();
       InputStream is = fileLookup.lookupFile("mapreduce/macbeth.txt", getClass().getClassLoader());
 
       BufferedReader br = new BufferedReader(new InputStreamReader(is));
