+++ demos/gui/src/main/java/org/infinispan/demo/InfinispanDemo.java
@@ -12,6 +12,7 @@ import org.infinispan.notifications.cachemanagerlistener.annotation.*;
 import org.infinispan.notifications.cachemanagerlistener.event.*;
 import org.infinispan.remoting.transport.Address;
 import org.infinispan.commons.util.FileLookup;
+import org.infinispan.commons.util.FileLookupFactory;
 import org.infinispan.commons.util.LegacyKeySupportSystemProperties;
 import org.infinispan.commons.util.Util;
 import org.infinispan.util.logging.Log;
@@ -353,7 +354,7 @@ public class InfinispanDemo {
          @Override
          public void run() {
             try {
-               URL resource = new FileLookup().lookupFileLocation(cacheConfigFile, getClass().getClassLoader());
+               URL resource = FileLookupFactory.newInstance().lookupFileLocation(cacheConfigFile, getClass().getClassLoader());
                if (resource == null) resource = new URL(cacheConfigFile);
 
                if (cacheManager == null) {
