+++ osgi/src/main/java/org/infinispan/osgi/InfinispanEmbeddedServiceFactory.java
@@ -9,9 +9,9 @@ import java.util.Hashtable;
 import java.util.Map;
 import java.util.Set;
 
+import org.infinispan.commons.util.FileLookupFactory;
 import org.infinispan.manager.DefaultCacheManager;
 import org.infinispan.manager.EmbeddedCacheManager;
-import org.infinispan.commons.util.FileLookup;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.cm.ConfigurationException;
@@ -50,7 +50,7 @@ public class InfinispanEmbeddedServiceFactory implements ManagedServiceFactory {
         }
 
         try {
-            URL configURL = new FileLookup().lookupFileLocation(config, Thread.currentThread().getContextClassLoader());
+            URL configURL = FileLookupFactory.newInstance().lookupFileLocation(config, Thread.currentThread().getContextClassLoader());
             if (configURL == null) {
                 throw new ConfigurationException(PROP_CONFIG, String.format("Failed to find the specified config '%s'.", config));
             }
